[connectivity] Fix parameter type mismatches.
To match the corresponding callback function prototype
(e.g. CANetworkPacketReceivedCallback and CAErrorHandleCallback), the
data length parameter for a few callback function implementations
should have been of type size_t instead uint32_t.  This addresses
several parameter type mismatch warnings on 64 bit platforms.

Reviewed-on: https://gerrit.iotivity.org/gerrit/14237
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ resource/csdk/connectivity/src/cainterfacecontroller.c
@@ -215,7 +215,7 @@ CAResult_t CASetAdapterRAInfo(const CARAInfo_t *caraInfo)
 #endif
 
 static void CAReceivedPacketCallback(const CASecureEndpoint_t *sep,
-                                     const void *data, uint32_t dataLen)
+                                     const void *data, size_t dataLen)
 {
     if (g_networkPacketReceivedCallback != NULL)
     {
@@ -265,7 +265,7 @@ static void CAConnectionChangedCallback(const CAEndpoint_t *info, bool isConnect
 #endif
 
 static void CAAdapterErrorHandleCallback(const CAEndpoint_t *endpoint,
-                                         const void *data, uint32_t dataLen,
+                                         const void *data, size_t dataLen,
                                          CAResult_t result)
 {
     OIC_LOG(DEBUG, TAG, "received error from adapter in interfacecontroller");
